services/product/requirements.txt
pymysql
fastapi
sqlalchemy
python-dotenv
fastapi_pagination
requests
python-multipart
python-magic
Wand
